refactor: :recycle: A2A-Refactor: Code agent refactor
