--- src/agents/code_agent/code_rag_agent_proxy.py
@@ -1,94 +1,0 @@
-"""Proxy for the Code RAG Agent with updated Guardrails v0.6.x usage."""
-
-import json, html
-from typing import Any
-from guardrails import Guard, OnFailAction
-from src.guardrails.validators.valid_html import IsHTMLField
-from src.guardrails.validators.valid_schema_json import ValidSchemaJson
-from src.logging_config import logger
-
-
-class CodeRAGAgentProxy:
-
-    OUTPUT_SCHEMA = {
-        "type": "object",
-        "properties": {
-            "html_code": {"type": "string"},
-            "generation_metadata": {
-                "type": "object",
-                "properties": {
-                    "model_used": {"type": "string"},
-                    "patterns_used": {"type": "integer"},
-                    "visual_components": {"type": "array", "items": {"type": "string"}},
-                    "custom_instructions": {"type": "string"},
-                    "timestamp": {"type": "string", "format": "date-time"},
-                },
-                "required": ["model_used", "patterns_used", "visual_components", "custom_instructions", "timestamp"],
-            },
-            "visual_analysis_summary": {
-                "type": "object",
-                "properties": {
-                    "components": {"type": "array", "items": {"type": "string"}},
-                    "layout": {"type": "string"},
-                    "style": {"type": "string"},
-                },
-                "required": ["components", "layout", "style"],
-            },
-        },
-        "required": ["html_code", "generation_metadata", "visual_analysis_summary"],
-    }
-
-    def __init__(self, agent):
-        logger.info("Inicializando CodeRAGAgentProxy")
-        self.agent = agent
-
-        self.output_guard = Guard.for_string(
-            validators=[
-                "json",  # valida que la salida sea JSON válido
-                ValidSchemaJson(json_schema=self.OUTPUT_SCHEMA, on_fail=OnFailAction.EXCEPTION),
-                IsHTMLField(on_fail=OnFailAction.EXCEPTION, property="html_code"),
-            ]
-        )
-
-        # TODO: Implementar sanitización si es necesario
-        self.sanitization_guard = None
-
-    def generate_code(self, patterns, visual_analysis, custom_instructions=""):
-        logger.info(
-            f"Llamada a generate_code con patterns={patterns}, "
-            f"visual_analysis={visual_analysis}, "
-            f"custom_instructions={custom_instructions}"
-        )
-        result = self.agent.generate_code(patterns, visual_analysis, custom_instructions)
-
-        try:
-            validated_output = self.output_guard.parse(json.dumps(result))
-        except Exception as e:
-            logger.error(f"Error durante la validación de salida: {e}")
-            validated_output = None
-            raise e
-
-        # Sanitización básica manual del HTML (antes era WebSanitization)
-        try:
-            html_code = result.get("html_code", "")
-            encoded_html = html.escape(html_code)
-            # Aquí podrías usar bleach.clean(encoded_html) si quieres más control
-        except Exception as e:
-            raise e
-
-        logger.info(f"Generación de código: {result}")
-        logger.info(f"Generación de código validada: {validated_output}")
-
-        return result
-
-    def get_rag_status(self):
-        logger.info("Llamada a get_rag_status")
-        return self.agent.get_rag_status()
-
-    def retrieve_patterns(self, visual_analysis: dict[str, Any], top_k: int = 5) -> list[tuple]:
-        logger.info(f"Llamada a retrieve_patterns con visual_analysis={visual_analysis}, " f"top_k={top_k}")
-        return self.agent.retrieve_patterns(visual_analysis, top_k)
-
-    def save_generated_code(self, code_result: dict[str, Any], filename: str = None) -> str:
-        logger.info(f"Llamada a save_generated_code con filename={filename}")
-        return self.agent.save_generated_code(code_result, filename)
