+++ src/agents/code_agent/src/example.prod.env
@@ -1,0 +1,1 @@
+OPENROUTER_API_KEY=your_openrouter_api_key_here
